Move the broker as a binary
