+++ src/lib.rs
@@ -1,10 +1,6 @@
-pub mod domobroker;
 pub mod domocache;
 pub mod domolibp2p;
 pub mod domopersistentstorage;
-pub mod restmessage;
 pub mod utils;
-pub mod webapimanager;
-pub mod websocketmessage;
 
 pub use libp2p::identity::Keypair;
